Implement test helper.
